refacotr: return an uint32_t as pixel colour wrapper functions; anti_alias new file
