+++ src/ray_bonus.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ray_bonus.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sqiu <sqiu@student.42vienna.com>           +#+  +:+       +#+        */
+/*   By: gwolf <gwolf@student.42vienna.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/27 12:05:20 by u413q             #+#    #+#             */
-/*   Updated: 2024/01/18 23:50:27 by sqiu             ###   ########.fr       */
+/*   Updated: 2024/01/20 15:35:19 by gwolf            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -65,28 +65,4 @@ t_colour	ft_background_colour(t_ray ray)
 	ray_colour.g = (1.0 - a) * white.g + a * blue.g;
 	ray_colour.b = (1.0 - a) * white.b + a * blue.b;
 	return (ray_colour);
-}
-
-t_ray	ft_create_sample_ray(t_vec3 pixel_centre, t_pixel_grid pixels, \
-	t_cam cam)
-{
-	t_vec3	pixel_sample;
-	t_ray	ray;
-
-	pixel_sample = ft_vec3_add(pixel_centre, ft_pixel_sample(pixels));
-	ray.origin = cam.centre;
-	ray.direction = ft_vec3_norm(ft_vec3_sub(pixel_sample, ray.origin));
-	ray.d = 1.0;
-	return (ray);
-}
-
-t_vec3	ft_pixel_sample(t_pixel_grid pixels)
-{
-	float	px;
-	float	py;
-
-	px = -0.5 + ft_random_float();
-	py = -0.5 + ft_random_float();
-	return (ft_vec3_add(ft_vec3_scale(pixels.delta_u, px), \
-		ft_vec3_scale(pixels.delta_v, py)));
 }
